requirements.txt
Django==1.9.4
django-nose==1.4.1
django-rest-swagger==2.1.2
djangorestframework==3.6.3
drf-extensions==0.3.1
jsonschema==2.6.0
MySQL-python==1.2.5